@@ -1,5 +1,6 @@
 package com.peakphysique.app.view
 
+import TrackingViewModel
 import androidx.compose.foundation.background
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Arrangement
@@ -16,42 +17,75 @@ import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.lazy.grid.GridCells
 import androidx.compose.foundation.lazy.grid.LazyVerticalGrid
 import androidx.compose.foundation.lazy.items
+import androidx.compose.foundation.shape.CircleShape
 import androidx.compose.material3.Button
+import androidx.compose.material3.ButtonDefaults
 import androidx.compose.material3.Card
 import androidx.compose.material3.CardDefaults
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.collectAsState
+import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.clip
 import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.text.TextStyle
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
+import androidx.lifecycle.viewmodel.compose.viewModel
 import androidx.navigation.NavController
 import com.peakphysique.app.controller.BottomNavBar
 import java.time.LocalDate
+import java.time.LocalDateTime
 import java.time.Month
 import java.time.YearMonth
 
-
 @Composable
-fun HistoryScreen(navController: NavController) {
-    // Existing state for date selection
+fun HistoryScreen(
+    navController: NavController,
+    viewModel: TrackingViewModel = viewModel()
+) {
     val selectedDay = remember { mutableStateOf(LocalDate.now().dayOfMonth) }
     val selectedMonth = remember { mutableStateOf(LocalDate.now().monthValue) }
     val selectedYear = remember { mutableStateOf(LocalDate.now().year) }
 
-    // New state for workout data
-    val workouts = remember {
-        mutableStateOf(listOf<WorkoutData>())
+    // Collect all workouts from the repository
+    val allWorkouts by viewModel.allWorkouts.collectAsState()
+
+    // Transform WorkoutWithSets to calendar data by extracting dates
+    val workoutDates = allWorkouts
+        .map { it.workout.date.toLocalDate() }  // Convert LocalDateTime to LocalDate
+        .filter { date ->  // Filter dates for the selected month/year
+            date.year == selectedYear.value &&
+                    date.monthValue == selectedMonth.value
+        }
+        .map { it.dayOfMonth }
+        .distinct()
+
+    // Filter workouts for selected date
+    val selectedDateWorkouts = allWorkouts.filter { workout ->
+        val workoutDate = workout.workout.date
+        workoutDate.year == selectedYear.value &&
+                workoutDate.monthValue == selectedMonth.value &&
+                workoutDate.dayOfMonth == selectedDay.value
     }
 
-    // Get the current date for the header
-    val currentDate = LocalDate.now()
-    val currentDateText = "${currentDate.year}-${currentDate.monthValue}-${currentDate.dayOfMonth}"
-    val selectedDateText = "${selectedYear.value}-${selectedMonth.value}-${selectedDay.value}"
+    // Transform WorkoutWithSets to WorkoutData for display
+    val workoutDisplayData = selectedDateWorkouts.flatMap { workoutWithSets ->
+        workoutWithSets.sets.groupBy { it.name }.map { (exerciseName, sets) ->
+            WorkoutData(
+                exerciseName = exerciseName,
+                sets = sets.size,
+                reps = sets.firstOrNull()?.reps?.toIntOrNull() ?: 0,
+                weight = sets.firstOrNull()?.weight?.toDoubleOrNull() ?: 0.0
+            )
+        }
+    }
 
     Column(
         modifier = Modifier
@@ -60,75 +94,119 @@ fun HistoryScreen(navController: NavController) {
         verticalArrangement = Arrangement.Top,
         horizontalAlignment = Alignment.CenterHorizontally
     ) {
-        // Display current and selected dates
-        Text(text = "Today's Date: $currentDateText")
-        Text(text = "Selected Date: $selectedDateText")
+        Spacer(modifier = Modifier.height(32.dp))
 
-        Spacer(modifier = Modifier.height(16.dp))
-
-        // Month navigation buttons
+        // Month navigation
         Row(
             modifier = Modifier.fillMaxWidth(),
             horizontalArrangement = Arrangement.SpaceBetween,
             verticalAlignment = Alignment.CenterVertically
         ) {
-            Button(onClick = {
-                if (selectedMonth.value == 1) {
-                    selectedMonth.value = 12
-                    selectedYear.value -= 1
-                } else {
-                    selectedMonth.value -= 1
-                }
-            }) {
+            Button(
+                onClick = {
+                    if (selectedMonth.value == 1) {
+                        selectedMonth.value = 12
+                        selectedYear.value -= 1
+                    } else {
+                        selectedMonth.value -= 1
+                    }
+                    selectedDay.value = 0
+                },
+                colors = ButtonDefaults.buttonColors(
+                    containerColor = Color(0xFF213455),
+                    contentColor = Color.White
+                )
+            ) {
                 Text("<")
             }
 
-            Text(
-                text = "${Month.of(selectedMonth.value).name.lowercase().capitalize()} - ${selectedYear.value}",
-            )
+            Column(
+                horizontalAlignment = Alignment.CenterHorizontally
+            ) {
+                Text(
+                    text = "${Month.of(selectedMonth.value).name.lowercase().capitalize()} - ${selectedYear.value}",
+                    fontWeight = FontWeight.Bold
+                )
 
-            Button(onClick = {
-                if (selectedMonth.value == 12) {
-                    selectedMonth.value = 1
-                    selectedYear.value += 1
-                } else {
-                    selectedMonth.value += 1
+                Button(
+                    onClick = {
+                        val today = LocalDate.now()
+                        selectedYear.value = today.year
+                        selectedMonth.value = today.monthValue
+                        selectedDay.value = today.dayOfMonth
+                    },
+                    colors = ButtonDefaults.buttonColors(
+                        containerColor = Color(0xFF213455),
+                        contentColor = Color.White
+                    ),
+                    modifier = Modifier.padding(top = 8.dp)
+                ) {
+                    Text("Today")
                 }
-            }) {
+            }
+
+            Button(
+                onClick = {
+                    if (selectedMonth.value == 12) {
+                        selectedMonth.value = 1
+                        selectedYear.value += 1
+                    } else {
+                        selectedMonth.value += 1
+                    }
+                    selectedDay.value = 0
+                },
+                colors = ButtonDefaults.buttonColors(
+                    containerColor = Color(0xFF213455),
+                    contentColor = Color.White
+                )
+            ) {
                 Text(">")
             }
         }
 
         Spacer(modifier = Modifier.height(16.dp))
 
-        // Display the calendar
+        // Calendar view
         CalendarView(
             year = selectedYear.value,
             month = selectedMonth.value,
-            workoutDates = listOf(2, 3), // Replace with actual workout dates
+            workoutDates = workoutDates,
             selectedDay = selectedDay.value,
-            currentDay = if (selectedYear.value == currentDate.year && selectedMonth.value == currentDate.monthValue) currentDate.dayOfMonth else -1,
+            currentDay = if (selectedYear.value == LocalDate.now().year &&
+                selectedMonth.value == LocalDate.now().monthValue)
+                LocalDate.now().dayOfMonth else -1,
             onDateClick = { day ->
                 selectedDay.value = day
-                // Here you would fetch workouts for the selected date
-                workouts.value = getWorkoutsForDate(selectedYear.value, selectedMonth.value, day)
             }
         )
 
         Spacer(modifier = Modifier.height(16.dp))
 
-        // Display workout cards for selected date
+        if (workoutDisplayData.isEmpty() && selectedDay.value > 0) {
+            Text(
+                text = "No workouts recorded for this date",
+                modifier = Modifier.padding(16.dp),
+                color = Color.Gray
+            )
+        }
+
+        // Workout cards
         LazyColumn(
             modifier = Modifier
                 .fillMaxWidth()
                 .weight(1f)
         ) {
-            items(workouts.value) { workout ->
+            items(workoutDisplayData) { workout ->
                 WorkoutCard(workout = workout)
             }
         }
     }
     BottomNavBar(navController)
+}
+
+// Add this extension function to support date conversions
+fun LocalDateTime.toLocalDate(): LocalDate {
+    return LocalDate.of(this.year, this.month, this.dayOfMonth)
 }
 
 @Composable
@@ -156,63 +234,16 @@ fun WorkoutCard(workout: WorkoutData) {
                 horizontalArrangement = Arrangement.SpaceBetween
             ) {
                 Column {
-                    Text(
-                        text = "Sets: ${workout.sets}",
-                    )
-                    Text(
-                        text = "Reps: ${workout.reps}",
-                    )
+                    Text(text = "Sets: ${workout.sets}")
+                    Text(text = "Reps: ${workout.reps}")
                 }
                 Column {
-                    Text(
-                        text = "Weight: ${workout.weight}lbs",
-                    )
+                    Text(text = "Weight: ${workout.weight}lbs")
                 }
             }
         }
     }
 }
-
-// Data class for workout information
-data class WorkoutData(
-    val exerciseName: String,
-    val sets: Int,
-    val reps: Int,
-    val weight: Double,
-)
-
-// Function to get workouts for a specific date
-fun getWorkoutsForDate(year: Int, month: Int, day: Int): List<WorkoutData> {
-    // Replace this with actual data
-    // This is just a sample implementation
-    return if (day % 2 == 0) {
-        listOf(
-            WorkoutData(
-                exerciseName = "Bench Press",
-                sets = 3,
-                reps = 10,
-                weight = 255.0,
-            ),
-            WorkoutData(
-                exerciseName = "Squats",
-                sets = 4,
-                reps = 8,
-                weight = 315.0,
-            )
-        )
-    } else {
-        listOf(
-            WorkoutData(
-                exerciseName = "Deadlift",
-                sets = 3,
-                reps = 5,
-                weight = 500.0,
-            )
-        )
-    }
-}
-
-
 
 @Composable
 fun WeekHeader() {
@@ -222,14 +253,15 @@ fun WeekHeader() {
         }
     }
 }
+
 @Composable
 fun CalendarView(
     year: Int,
     month: Int,
-    workoutDates: List<Int>,  // List of days with workouts
-    selectedDay: Int, // Day of the month selected by user
-    currentDay : Int, // Current day of the month
-    onDateClick: (Int) -> Unit // Callback for date selection
+    workoutDates: List<Int>,
+    selectedDay: Int,
+    currentDay: Int,
+    onDateClick: (Int) -> Unit
 ) {
     val daysInMonth = YearMonth.of(year, month).lengthOfMonth()
     val firstDayOfWeek = LocalDate.of(year, month, 1).dayOfWeek.value % 7
@@ -257,26 +289,43 @@ fun CalendarView(
         }
     }
 }
+
 @Composable
 fun DayCell(
     day: Int,
     isWorkoutDay: Boolean,
     isSelected: Boolean,
-    onClick: () -> Unit,
-    currentDay: Boolean
+    currentDay: Boolean,
+    onClick: () -> Unit
 ) {
     Box(
         modifier = Modifier
             .size(40.dp)
-            .clickable(onClick = onClick)
+            .padding(4.dp)
+            .clip(CircleShape)
             .background(when {
-                isSelected -> Color(0xFF003D6E) // Navy blue for selected date
-                isWorkoutDay -> Color(0x8A7CA8FF)      // Black for workout days
-                currentDay -> Color(0xFF969696) // Current day is gray
-                else -> Color.Transparent        // Transparent for empty days
-            }),
+                isSelected -> Color(0xFF003D6E)
+                isWorkoutDay -> Color(0x8A7CA8FF)
+                currentDay -> Color(0xFF969696)
+                else -> Color.Transparent
+            })
+            .clickable(onClick = onClick),
         contentAlignment = Alignment.Center
     ) {
-        Text(text = day.toString(), color = Color.Black)
+        Text(
+            text = day.toString(),
+            color = if (isSelected) Color.White else Color.Black,
+            style = TextStyle(
+                fontSize = 14.sp,
+                fontWeight = if (currentDay) FontWeight.Bold else FontWeight.Normal
+            )
+        )
     }
 }
+
+data class WorkoutData(
+    val exerciseName: String,
+    val sets: Int,
+    val reps: Int,
+    val weight: Double,
+)
