@@ -1,7 +1,11 @@
 package com.peakphysique.app.view
 
+import androidx.compose.foundation.Image
+import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.*
 import androidx.compose.foundation.rememberScrollState
+import androidx.compose.foundation.shape.CircleShape
+import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.foundation.verticalScroll
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.*
@@ -9,176 +13,202 @@ import androidx.compose.material3.*
 import androidx.compose.runtime.*
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.clip
+import androidx.compose.ui.graphics.Brush
+import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.graphics.vector.ImageVector
+import androidx.compose.ui.layout.ContentScale
+import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.unit.dp
 import androidx.navigation.NavController
 import com.peakphysique.app.controller.BottomNavBar
-
-// TODO: Pull user data from database
+import com.peakphysique.app.viewmodel.AccountViewModel
+import androidx.lifecycle.viewmodel.compose.viewModel
+import com.peakphysique.app.R
 
 @Composable
-fun AccountScreen(navController: NavController,
-                  onEditProfile: () -> Unit = {},
-                  onViewProgress: () -> Unit = {},
-                  onViewWorkoutHistory: () -> Unit = {},
-                  onViewAchievements: () -> Unit = {},
-                  onSettings: () -> Unit = {},
-                  onLogout: () -> Unit = {}
+fun AccountScreen(
+    navController: NavController,
+    viewModel: AccountViewModel = viewModel(),
 ) {
+    val totalWorkouts by viewModel.totalWorkouts.collectAsState()
+    val recordsBroken by viewModel.recordsBroken.collectAsState()
+    val monthsActive by viewModel.monthsActive.collectAsState()
+    val displayName by viewModel.displayName.collectAsState()
+
     val scrollState = rememberScrollState()
 
-    Column(
+    Box(
         modifier = Modifier
             .fillMaxSize()
-            .verticalScroll(scrollState)
-            .padding(16.dp),
-        horizontalAlignment = Alignment.CenterHorizontally
+            .background(MaterialTheme.colorScheme.background)
     ) {
-        // Profile Section
-        Card(
+        Column(
             modifier = Modifier
-                .fillMaxWidth(0.9f)
-                .padding(bottom = 16.dp)
+                .fillMaxSize()
+                .verticalScroll(scrollState)
+                .padding(top = 45.dp, start = 16.dp, end = 16.dp),
+            horizontalAlignment = Alignment.CenterHorizontally
         ) {
-            Box(
-                modifier = Modifier.fillMaxWidth(),
-                contentAlignment = Alignment.Center
+            Card(
+                modifier = Modifier
+                    .fillMaxWidth()
+                    .padding(bottom = 24.dp),
+                shape = RoundedCornerShape(24.dp),
             ) {
-                Column(
-                    modifier = Modifier.padding(16.dp),
-                    horizontalAlignment = Alignment.CenterHorizontally,
-                    verticalArrangement = Arrangement.Center
+                Box(
+                    modifier = Modifier
+                        .fillMaxWidth()
+                        .background(
+                            Brush.verticalGradient(
+                                colors = listOf(
+                                    MaterialTheme.colorScheme.primary.copy(alpha = 0.8f),
+                                    MaterialTheme.colorScheme.primary.copy(alpha = 0.4f)
+                                )
+                            )
+                        )
                 ) {
-                    // Profile Picture
-                    Surface(
-                        modifier = Modifier
-                            .padding(top = 34.dp)
-                            .size(100.dp),
-                        shape = MaterialTheme.shapes.medium,
-                        tonalElevation = 1.dp
+                    Column(
+                        modifier = Modifier.padding(24.dp),
+                        horizontalAlignment = Alignment.CenterHorizontally
                     ) {
-                        Icon(
-                            imageVector = Icons.Default.Person,
-                            contentDescription = "Profile Picture",
+                        // Profile Picture with Border
+                        Surface(
                             modifier = Modifier
-                                .padding(16.dp)
-                                .size(48.dp)
+                                .size(120.dp)
+                                .clip(CircleShape),
+                            color = MaterialTheme.colorScheme.surface,
+                            tonalElevation = 4.dp
+                        ) {
+                            Image(
+                                painter = painterResource(id = R.drawable.logo_),
+                                contentDescription = "Profile Picture",
+                                modifier = Modifier
+                                    .padding(8.dp)
+                                    .fillMaxSize(),
+                                contentScale = ContentScale.Fit
+                            )
+                        }
+
+                        Spacer(modifier = Modifier.height(16.dp))
+
+                        Text(
+                            text = displayName,
+                            style = MaterialTheme.typography.headlineMedium,
+                            fontWeight = FontWeight.Bold,
+                            color = MaterialTheme.colorScheme.onPrimary
+                        )
+
+                        Text(
+                            text = "Member since: Jan 2024",
+                            style = MaterialTheme.typography.bodyMedium,
+                            color = MaterialTheme.colorScheme.onPrimary.copy(alpha = 0.8f)
                         )
                     }
-
-                    Spacer(modifier = Modifier.height(16.dp))
-
-                    Text(
-                        text = "John Doe",
-                        style = MaterialTheme.typography.headlineSmall,
-                        fontWeight = FontWeight.Bold
-                    )
-
-                    Text(
-                        text = "Member since: Jan 2024",
-                        style = MaterialTheme.typography.bodyMedium,
-                        color = MaterialTheme.colorScheme.onSurfaceVariant
-                    )
-
-                    Button(
-                        onClick = onEditProfile,
-                        modifier = Modifier.padding(top = 8.dp)
-                    ) {
-                        Text("Edit Profile")
-                    }
                 }
             }
-        }
 
-        // Stats Section
-        Card(
-            modifier = Modifier
-                .fillMaxWidth()
-                .padding(bottom = 16.dp)
-        ) {
-            Column(
-                modifier = Modifier.padding(16.dp)
+            // Stats Section with Enhanced Cards
+            Row(
+                modifier = Modifier
+                    .fillMaxWidth()
+                    .padding(bottom = 24.dp),
+                horizontalArrangement = Arrangement.spacedBy(12.dp)
             ) {
-                Text(
-                    text = "Quick Stats",
-                    style = MaterialTheme.typography.titleMedium,
-                    fontWeight = FontWeight.Bold
+                StatCard(
+                    modifier = Modifier.weight(1f),
+                    icon = Icons.Default.Build,
+                    number = totalWorkouts.toString(),
+                    label = "Workouts"
                 )
+                StatCard(
+                    modifier = Modifier.weight(1f),
+                    icon = Icons.Default.Star,
+                    number = recordsBroken.toString(),
+                    label = "Records"
+                )
+                StatCard(
+                    modifier = Modifier.weight(1f),
+                    icon = Icons.Default.DateRange,
+                    number = monthsActive.toString(),
+                    label = "Months"
+                )
+            }
 
-                Spacer(modifier = Modifier.height(8.dp))
-
-                Row(
-                    modifier = Modifier.fillMaxWidth(),
-                    horizontalArrangement = Arrangement.SpaceEvenly
+            // Menu Items with Enhanced Styling
+            Card(
+                modifier = Modifier.fillMaxWidth(),
+                shape = RoundedCornerShape(16.dp)
+            ) {
+                Column(
+                    modifier = Modifier.padding(8.dp)
                 ) {
-                    StatItem(number = "98", label = "Workouts")
-                    StatItem(number = "12", label = "Records Broken")
-                    StatItem(number = "6", label = "Months Active")
+                    MenuListItem(
+                        icon = Icons.Default.FavoriteBorder,
+                        title = "View Progress",
+                        subtitle = "Track your fitness journey",
+                        onClick = { navController.navigate("progress_screen") }
+                    )
+
+                    MenuListItem(
+                        icon = Icons.Default.Add,
+                        title = "Log Weight",
+                        subtitle = "Update your current weight",
+                        onClick = { navController.navigate("log_weight_screen") }
+                    )
+
+                    MenuListItem(
+                        icon = Icons.Default.Settings,
+                        title = "Settings",
+                        subtitle = "Customize your experience",
+                        onClick = { navController.navigate("settings_screen") }
+                    )
                 }
             }
+
+            Spacer(modifier = Modifier.height(80.dp))
         }
 
-        // Menu Items
-        MenuListItem(
-            icon = Icons.Default.FavoriteBorder,
-            title = "View Progress",
-            onClick = { navController.navigate("progress_screen") }
+        BottomNavBar(
+            navController = navController
         )
-
-        MenuListItem(
-            icon = Icons.Default.DateRange,
-            title = "Workout History",
-            onClick = onViewWorkoutHistory
-        )
-
-        MenuListItem(
-            icon = Icons.Default.Star,
-            title = "Achievements",
-            onClick = onViewAchievements
-        )
-
-        MenuListItem(
-            icon = Icons.Default.Settings,
-            title = "Settings",
-            onClick = onSettings
-        )
-
-        Spacer(modifier = Modifier.height(16.dp))
-
-        OutlinedButton(
-            onClick = onLogout,
-            modifier = Modifier.fillMaxWidth()
-        ) {
-            Icon(
-                imageVector = Icons.Default.ExitToApp,
-                contentDescription = null,
-                modifier = Modifier.padding(end = 8.dp)
-            )
-            Text("Logout")
-        }
     }
-    BottomNavBar(navController = navController)
 }
 
 @Composable
-private fun StatItem(
+private fun StatCard(
+    modifier: Modifier = Modifier,
+    icon: ImageVector,
     number: String,
     label: String
 ) {
-    Column(
-        horizontalAlignment = Alignment.CenterHorizontally
+    Card(
+        modifier = modifier,
+        shape = RoundedCornerShape(16.dp)
     ) {
-        Text(
-            text = number,
-            style = MaterialTheme.typography.titleLarge,
-            fontWeight = FontWeight.Bold
-        )
-        Text(
-            text = label,
-            style = MaterialTheme.typography.bodySmall,
-            color = MaterialTheme.colorScheme.onSurfaceVariant
-        )
+        Column(
+            modifier = Modifier.padding(16.dp),
+            horizontalAlignment = Alignment.CenterHorizontally
+        ) {
+            Icon(
+                imageVector = icon,
+                contentDescription = null,
+                tint = MaterialTheme.colorScheme.primary,
+                modifier = Modifier.size(24.dp)
+            )
+            Spacer(modifier = Modifier.height(8.dp))
+            Text(
+                text = number,
+                style = MaterialTheme.typography.titleLarge,
+                fontWeight = FontWeight.Bold
+            )
+            Text(
+                text = label,
+                style = MaterialTheme.typography.bodySmall,
+                color = MaterialTheme.colorScheme.onSurfaceVariant
+            )
+        }
     }
 }
 
@@ -186,36 +216,53 @@ private fun StatItem(
 private fun MenuListItem(
     icon: ImageVector,
     title: String,
+    subtitle: String,
     onClick: () -> Unit
 ) {
     Surface(
         onClick = onClick,
-        modifier = Modifier
-            .fillMaxWidth()
-            .padding(vertical = 4.dp)
+        color = Color.Transparent
     ) {
         Row(
             modifier = Modifier
+                .fillMaxWidth()
                 .padding(16.dp),
             verticalAlignment = Alignment.CenterVertically
         ) {
-            Icon(
-                imageVector = icon,
-                contentDescription = null,
-                tint = MaterialTheme.colorScheme.primary
-            )
+            Surface(
+                shape = CircleShape,
+                color = MaterialTheme.colorScheme.primary.copy(alpha = 0.1f),
+                modifier = Modifier.size(40.dp)
+            ) {
+                Box(
+                    contentAlignment = Alignment.Center,
+                    modifier = Modifier.fillMaxSize()
+                ) {
+                    Icon(
+                        imageVector = icon,
+                        contentDescription = null,
+                        tint = MaterialTheme.colorScheme.primary
+                    )
+                }
+            }
 
             Spacer(modifier = Modifier.width(16.dp))
 
-            Text(
-                text = title,
-                style = MaterialTheme.typography.bodyLarge
-            )
-
-            Spacer(modifier = Modifier.weight(1f))
+            Column(modifier = Modifier.weight(1f)) {
+                Text(
+                    text = title,
+                    style = MaterialTheme.typography.bodyLarge,
+                    fontWeight = FontWeight.Medium
+                )
+                Text(
+                    text = subtitle,
+                    style = MaterialTheme.typography.bodySmall,
+                    color = MaterialTheme.colorScheme.onSurfaceVariant
+                )
+            }
 
             Icon(
-                imageVector = Icons.Default.KeyboardArrowDown,
+                imageVector = Icons.Default.ArrowDropDown,
                 contentDescription = null,
                 tint = MaterialTheme.colorScheme.onSurfaceVariant
             )
