@@ -1,47 +1,79 @@
 package com.peakphysique.app.view
 
+import android.app.Application
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.*
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.lazy.items
 import androidx.compose.foundation.shape.CircleShape
 import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.Close
-import androidx.compose.material.icons.filled.Favorite
-import androidx.compose.material.icons.filled.Home
-import androidx.compose.material.icons.filled.Star
 import androidx.compose.material3.Card
 import androidx.compose.material3.CardDefaults
 import androidx.compose.material3.Icon
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.collectAsState
+import androidx.compose.runtime.getValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.clip
 import androidx.compose.ui.graphics.Brush
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.graphics.vector.ImageVector
+import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import androidx.navigation.NavController
 import com.peakphysique.app.controller.BottomNavBar
+import com.peakphysique.app.viewmodel.FeedViewModel
+import androidx.lifecycle.viewmodel.compose.viewModel
+import com.peakphysique.app.ui.theme.Buttons
+import com.peakphysique.app.viewmodel.FeedViewModelFactory
 
 @Composable
-fun FeedScreen(navController: NavController, modifier: Modifier = Modifier) {
+fun FeedScreen(
+    navController: NavController,
+    viewModel: FeedViewModel = viewModel(
+        factory = FeedViewModelFactory.provide(LocalContext.current.applicationContext as Application)
+    ),
+    modifier: Modifier = Modifier
+) {
+    val feedEntries by viewModel.feedEntries.collectAsState()
+
     Column(
         modifier = modifier
             .fillMaxWidth()
             .padding(horizontal = 16.dp, vertical = 44.dp)
     ) {
-        LazyColumn(
-            modifier = Modifier
-                .fillMaxSize()
-                .padding(bottom = 60.dp)
-        ) {
-            items(feedData) { entry ->
-                FeedEntry(type = entry.type, achievement = entry.achievement, time = entry.time, icon = entry.icon)
+        if (feedEntries.isEmpty()) {
+            // Modified with LazyColumn below to if else to allow empty feed to take text
+            Box(
+                modifier = Modifier.fillMaxSize(),
+                contentAlignment = Alignment.Center
+            ) {
+                Text(
+                    text = "      No Achievements?\n\n \uD83C\uDFCB\uFE0F\u200D♂\uFE0F Lets Get Started! \uD83C\uDFCB\uFE0F\u200D♀\uFE0F",
+                    fontSize = 30.sp,
+                    fontWeight = FontWeight.Bold,
+                    color = Buttons,
+                    modifier = Modifier.offset(y = (-70).dp)
+                )
+            }
+        } else {
+            LazyColumn(
+                modifier = Modifier
+                    .fillMaxSize()
+                    .padding(bottom = 60.dp)
+            ) {
+                items(feedEntries) { entry ->
+                    FeedEntry(
+                        type = entry.type,
+                        achievement = entry.achievement,
+                        time = entry.time,
+                        icon = entry.icon
+                    )
+                }
             }
         }
     }
@@ -51,29 +83,6 @@ fun FeedScreen(navController: NavController, modifier: Modifier = Modifier) {
 
 data class FeedEntryData(val type: String, val achievement: String, val time: String, val icon: ImageVector)
 
-// Hardcoded feed data for now until we incorporate a database
-val feedData = listOf(
-    FeedEntryData("New Bench Press PR", "225lbs Wow! That's a 20% increase from last month!", "2 hours ago", Icons.Default.Favorite),
-    FeedEntryData("Get out there and try again", "You deadlifted 205 this week, That's a 10% decrease from last month.", "3 hours ago", Icons.Default.Close),
-    FeedEntryData("New Achievement", "Completed 5 workouts this week", "4 hours ago", Icons.Default.Star),
-    FeedEntryData("New Achievement", "Completed 10 workouts this month", "10 hours ago", Icons.Default.Star),
-    FeedEntryData("You're on fire!", "You've worked out 5 days in a row", "1 day ago", Icons.Default.Star),
-    FeedEntryData("New Bench Press PR", "225lbs Wow! That's a 20% increase from last month!", "2 hours ago", Icons.Default.Favorite),
-    FeedEntryData("Get out there and try again", "You deadlifted 205 this week, That's a 10% decrease from last month.", "3 hours ago", Icons.Default.Close),
-    FeedEntryData("New Achievement", "Completed 5 workouts this week", "4 hours ago", Icons.Default.Star),
-    FeedEntryData("New Achievement", "Completed 10 workouts this month", "10 hours ago", Icons.Default.Star),
-    FeedEntryData("You're on fire!", "You've worked out 5 days in a row", "1 day ago", Icons.Default.Star),
-    FeedEntryData("New Bench Press PR", "225lbs Wow! That's a 20% increase from last month!", "2 hours ago", Icons.Default.Favorite),
-    FeedEntryData("Get out there and try again", "You deadlifted 205 this week, That's a 10% decrease from last month.", "3 hours ago", Icons.Default.Close),
-    FeedEntryData("New Achievement", "Completed 5 workouts this week", "4 hours ago", Icons.Default.Star),
-    FeedEntryData("New Achievement", "Completed 10 workouts this month", "10 hours ago", Icons.Default.Star),
-    FeedEntryData("You're on fire!", "You've worked out 5 days in a row", "1 day ago", Icons.Default.Star),
-    FeedEntryData("New Bench Press PR", "225lbs Wow! That's a 20% increase from last month!", "2 hours ago", Icons.Default.Favorite),
-    FeedEntryData("Get out there and try again", "You deadlifted 205 this week, That's a 10% decrease from last month.", "3 hours ago", Icons.Default.Close),
-    FeedEntryData("New Achievement", "Completed 5 workouts this week", "4 hours ago", Icons.Default.Star),
-    FeedEntryData("New Achievement", "Completed 10 workouts this month", "10 hours ago", Icons.Default.Star),
-    FeedEntryData("You're on fire!", "You've worked out 5 days in a row", "1 day ago", Icons.Default.Star),
-)
 
 @Composable
 fun FeedEntry(type: String, achievement: String, time: String, icon: ImageVector, modifier: Modifier = Modifier) {
